+++ lib/rspec/que/queue_up/queued_something.rb
@@ -1,0 +1,22 @@
+module RSpec
+  module Que
+    module Matchers
+      class QueueUp
+        class QueuedSomething
+          def matches?(_job)
+            true
+          end
+
+          def desc
+            "a job"
+          end
+
+          def failed_msg(_last_found)
+            "nothing"
+          end
+        end
+      end
+    end
+  end
+end
+
